refactor: extract TitleVersion component
